Add listing of stale items.
The stale items are sampled from all areaDetector repositories.

In order to achieve this, it was necessary to increase the amount of
issues and PRs that could be returned from gh, since otherwise they
would be limited to 30, and could miss one of the stale issues. 1000 is
used as a stand-in for "infinite".